+++ packages/core/src/vcdm/encoding/rlp/index.ts
@@ -1,10 +1,10 @@
-import { RLPBase } from './rlp';
 import { RLPProfiles } from './kind';
+import { RLPBase } from './rlp';
 
-export * from './rlp';
 export * from './kind';
+export * from './kind/helpers';
+export * from './rlp';
 export * from './types.d';
-export * from './helpers';
 
 /**
  * The `RLP_CODER` object, encapsulating all functionalities related to Recursive Length Prefix (RLP_CODER) encoding.
