+++ packages/core/src/vcdm/encoding/rlp/kind/helpers/fixedhexblobkind.ts
