+++ packages/core/src/vcdm/encoding/rlp/kind/helpers/numerickind.ts
@@ -1,6 +1,6 @@
 import { InvalidRLP } from '@vechain/sdk-errors';
-import { FPN, Hex, HexUInt } from '../../../../vcdm';
-import { type RLPInput } from '../types';
+import { FPN, Hex, HexUInt } from '../../../..';
+import { type RLPInput } from '../../types';
 
 /**
  * Validates and converts the input data to a BigInt.
