+++ packages/core/src/vcdm/encoding/rlp/kind/hexblobkinds/fixedhexblob.ts
@@ -1,8 +1,8 @@
+import { type BufferOutput, type DataOutput, type RLPInput } from '../../types';
 import {
     assertFixedHexBlobKindBuffer,
     assertFixedHexBlobKindData
-} from '../../helpers';
-import { type BufferOutput, type DataOutput, type RLPInput } from '../../types';
+} from '../helpers';
 import { HexBlobKind } from './hexblob';
 
 /**
